Create add_basic.f90
cannot say this is the most elegant way and perhaps not the best FORTRAN coding practices. known issues involve passing to assumed-size arrays in FORTRAN. currently gfortran does not support assumed-shape arrays so assumed-size arrays can be used but can easily be an issue if accessing out of bound indices.